[NFC] Fix some types on loops for getNumPorts.  Thanks to @Tang-Haojin.  May fix windows build.

+++ lib/Dialect/FIRRTL/Transforms/Dedup.cpp
@@ -1400,7 +1400,7 @@ void fixupAllModules(InstanceGraph &instanceGraph) {
       ImplicitLocOpBuilder builder(inst.getLoc(), inst->getContext());
       builder.setInsertionPointAfter(inst);
       auto e = getNumPorts(module);
-      for (unsigned i = 0; i < e; ++i) {
+      for (size_t i = 0; i < e; ++i) {
         auto result = inst.getResult(i);
         auto newType = module.getPortType(i);
         auto oldType = result.getType();
